step 3 - 클래스에 맞는 메소드 이동 Racing.resultView -> OutputView GenerateRandom.moveCarForward -> Racing.moveCarForward
