+++ src/main/java/racingcar/view/OutputView.java
@@ -3,6 +3,8 @@ package racingcar.view;
 import racingcar.car.Car;
 import java.util.List;
 import static java.lang.System.*;
+import static racingcar.racing.Racing.moveCarForward;
+import static racingcar.random.GenerateRandom.randomForward;
 
 public class OutputView {
     public static void printCarStatus(List<Car> cars) {
@@ -13,5 +15,14 @@ public class OutputView {
             }
             out.println(track);
         });
+    }
+
+    public static void resultView(List<Car> cars, int tryCount) {
+        out.println("실행 결과");
+        for(int i=0; i < tryCount; i++) {
+            cars = moveCarForward(cars, randomForward());
+            printCarStatus(cars);
+            out.println();
+        }
     }
 }
